expand signature with aux
